requirements.txt
gradio
openai
numpy
loguru
python-dotenv
markdown
tenacity